Implement Consonance Order coloring, manually revert legend location
Coloring was modularized into 3 coloring functions:
colorByLinearChromatic, ColorByConsonanceOrder, and
ColorByCircleOfFifths (not yet implemented)

Legend was manually reverted to the bottom of the FFT section of the
screen, because the 'git stash pop' failed to apply the stashed changes
after attempting to revert the buggy commit.

The max HSB hue value was set to 300 (cf. 360) to end the colors at
violet and omit magenta, since magenta looks too close to red.

Added a checkbox for decoloring accidentals (not yet implemented)

TODO:
     - Enums types in switch are just string comparisons and should be
       replaced with integer comparsions.
     - Implement circle of fifths coloring
     - Implement decoloring
     - Map volume amplitude  to stay within its section of the screen.
     - Map the scaled FFT bars to stay within their section of the
       screen.